Update DBFT format. Update RPC UT cases
