+++ src/DBFTPlugin/Consensus/ConsensusService.cs
@@ -1,15 +1,11 @@
 using Akka.Actor;
-using Neo.Cryptography;
 using Neo.IO;
 using Neo.Ledger;
 using Neo.Network.P2P;
 using Neo.Network.P2P.Payloads;
-using Neo.SmartContract;
-using Neo.SmartContract.Native;
 using Neo.Wallets;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using static Neo.Ledger.Blockchain;
 
@@ -48,9 +44,7 @@ namespace Neo.Consensus
         private readonly NeoSystem neoSystem;
 
         public ConsensusService(NeoSystem neoSystem, Settings settings, Wallet wallet)
-            : this(neoSystem, settings, new ConsensusContext(neoSystem, settings, wallet))
-        {
-        }
+            : this(neoSystem, settings, new ConsensusContext(neoSystem, settings, wallet)){}
 
         internal ConsensusService(NeoSystem neoSystem, Settings settings, ConsensusContext context)
         {
@@ -64,38 +58,12 @@ namespace Neo.Consensus
             Context.System.EventStream.Subscribe(Self, typeof(Blockchain.RelayResult));
         }
 
-        private bool AddTransaction(Transaction tx, bool verify)
+        private void OnPersistCompleted(Block block)
         {
-            if (verify)
-            {
-                VerifyResult result = tx.Verify(neoSystem.Settings, context.Snapshot, context.VerificationContext);
-                if (result != VerifyResult.Succeed)
-                {
-                    Log($"Rejected tx: {tx.Hash}, {result}{Environment.NewLine}{tx.ToArray().ToHexString()}", LogLevel.Warning);
-                    RequestChangeView(result == VerifyResult.PolicyFail ? ChangeViewReason.TxRejectedByPolicy : ChangeViewReason.TxInvalid);
-                    return false;
-                }
-            }
-            context.Transactions[tx.Hash] = tx;
-            context.VerificationContext.AddTransaction(tx);
-            return CheckPrepareResponse();
+            Log($"Persisted {nameof(Block)}: height={block.Index} hash={block.Hash} tx={block.Transactions.Length} nonce={block.Nonce}");
+            knownHashes.Clear();
+            InitializeConsensus(0);
         }
-
-       
-
-        private void ChangeTimer(TimeSpan delay)
-        {
-            clock_started = TimeProvider.Current.UtcNow;
-            expected_delay = delay;
-            timer_token.CancelIfNotNull();
-            timer_token = Context.System.Scheduler.ScheduleTellOnceCancelable(delay, Self, new Timer
-            {
-                Height = context.Block.Index,
-                ViewNumber = context.ViewNumber
-            }, ActorRefs.NoSender);
-        }
-
-       
 
         private void InitializeConsensus(byte viewNumber)
         {
@@ -130,31 +98,6 @@ namespace Neo.Consensus
             }
         }
 
-        private static void Log(string message, LogLevel level = LogLevel.Info)
-        {
-            Utility.Log(nameof(ConsensusService), level, message);
-        }
-
-       
-
-        // this function increases existing timer (never decreases) with a value proportional to `maxDelayInBlockTimes`*`Blockchain.MillisecondsPerBlock`
-        private void ExtendTimerByFactor(int maxDelayInBlockTimes)
-        {
-            TimeSpan nextDelay = expected_delay - (TimeProvider.Current.UtcNow - clock_started) + TimeSpan.FromMilliseconds(maxDelayInBlockTimes * neoSystem.Settings.MillisecondsPerBlock / (double)context.M);
-            if (!context.WatchOnly && !context.ViewChanging && !context.CommitSent && (nextDelay > TimeSpan.Zero))
-                ChangeTimer(nextDelay);
-        }
-
-       
-        private void OnPersistCompleted(Block block)
-        {
-            Log($"Persisted {nameof(Block)}: height={block.Index} hash={block.Hash} tx={block.Transactions.Length} nonce={block.Nonce}");
-            knownHashes.Clear();
-            InitializeConsensus(0);
-        }
-
-        
-
         protected override void OnReceive(object message)
         {
             if (message is Start)
@@ -182,64 +125,6 @@ namespace Neo.Consensus
                         break;
                 }
             }
-        }
-
-        private void OnConsensusPayload(ExtensiblePayload payload)
-        {
-            if (context.BlockSent) return;
-            ConsensusMessage message;
-            try
-            {
-                message = context.GetMessage(payload);
-            }
-            catch (FormatException)
-            {
-                return;
-            }
-            catch (IOException)
-            {
-                return;
-            }
-            if (!message.Verify(neoSystem.Settings)) return;
-            if (message.BlockIndex != context.Block.Index)
-            {
-                if (context.Block.Index < message.BlockIndex)
-                {
-                    Log($"Chain is behind: expected={message.BlockIndex} current={context.Block.Index - 1}", LogLevel.Warning);
-                }
-                return;
-            }
-            if (message.ValidatorIndex >= context.Validators.Length) return;
-            if (payload.Sender != Contract.CreateSignatureRedeemScript(context.Validators[message.ValidatorIndex]).ToScriptHash()) return;
-            context.LastSeenMessage[context.Validators[message.ValidatorIndex]] = message.BlockIndex;
-            switch (message)
-            {
-                case ChangeView view:
-                    OnChangeViewReceived(payload, view);
-                    break;
-                case PrepareRequest request:
-                    OnPrepareRequestReceived(payload, request);
-                    break;
-                case PrepareResponse response:
-                    OnPrepareResponseReceived(payload, response);
-                    break;
-                case Commit commit:
-                    OnCommitReceived(payload, commit);
-                    break;
-                case RecoveryRequest request:
-                    OnRecoveryRequestReceived(payload, request);
-                    break;
-                case RecoveryMessage recovery:
-                    OnRecoveryMessageReceived(recovery);
-                    break;
-            }
-        }
-
-
-        private void RequestRecovery()
-        {
-            Log($"Sending {nameof(RecoveryRequest)}: height={context.Block.Index} view={context.ViewNumber} nc={context.CountCommitted} nf={context.CountFailed}");
-            localNode.Tell(new LocalNode.SendDirectly { Inventory = context.MakeRecoveryRequest() });
         }
 
         private void OnStart()
@@ -298,27 +183,26 @@ namespace Neo.Consensus
             }
         }
 
-        private void OnTransaction(Transaction transaction)
+        private void SendPrepareRequest()
         {
-            if (!context.IsBackup || context.NotAcceptingPayloadsDueToViewChanging || !context.RequestSentOrReceived || context.ResponseSent || context.BlockSent)
-                return;
-            if (context.Transactions.ContainsKey(transaction.Hash)) return;
-            if (!context.TransactionHashes.Contains(transaction.Hash)) return;
-            AddTransaction(transaction, true);
+            Log($"Sending {nameof(PrepareRequest)}: height={context.Block.Index} view={context.ViewNumber}");
+            localNode.Tell(new LocalNode.SendDirectly { Inventory = context.MakePrepareRequest() });
+
+            if (context.Validators.Length == 1)
+                CheckPreparations();
+
+            if (context.TransactionHashes.Length > 0)
+            {
+                foreach (InvPayload payload in InvPayload.CreateGroup(InventoryType.TX, context.TransactionHashes))
+                    localNode.Tell(Message.Create(MessageCommand.Inv, payload));
+            }
+            ChangeTimer(TimeSpan.FromMilliseconds((neoSystem.Settings.MillisecondsPerBlock << (context.ViewNumber + 1)) - (context.ViewNumber == 0 ? neoSystem.Settings.MillisecondsPerBlock : 0)));
         }
 
-        protected override void PostStop()
+        private void RequestRecovery()
         {
-            Log("OnStop");
-            started = false;
-            Context.System.EventStream.Unsubscribe(Self);
-            context.Dispose();
-            base.PostStop();
-        }
-
-        public static Props Props(NeoSystem neoSystem, Settings dbftSettings, Wallet wallet)
-        {
-            return Akka.Actor.Props.Create(() => new ConsensusService(neoSystem, dbftSettings, wallet));
+            Log($"Sending {nameof(RecoveryRequest)}: height={context.Block.Index} view={context.ViewNumber} nc={context.CountCommitted} nf={context.CountFailed}");
+            localNode.Tell(new LocalNode.SendDirectly { Inventory = context.MakeRecoveryRequest() });
         }
 
         private void RequestChangeView(ChangeViewReason reason)
@@ -350,20 +234,69 @@ namespace Neo.Consensus
             return true;
         }
 
-        private void SendPrepareRequest()
+        private void OnTransaction(Transaction transaction)
         {
-            Log($"Sending {nameof(PrepareRequest)}: height={context.Block.Index} view={context.ViewNumber}");
-            localNode.Tell(new LocalNode.SendDirectly { Inventory = context.MakePrepareRequest() });
+            if (!context.IsBackup || context.NotAcceptingPayloadsDueToViewChanging || !context.RequestSentOrReceived || context.ResponseSent || context.BlockSent)
+                return;
+            if (context.Transactions.ContainsKey(transaction.Hash)) return;
+            if (!context.TransactionHashes.Contains(transaction.Hash)) return;
+            AddTransaction(transaction, true);
+        }
 
-            if (context.Validators.Length == 1)
-                CheckPreparations();
-
-            if (context.TransactionHashes.Length > 0)
+        private bool AddTransaction(Transaction tx, bool verify)
+        {
+            if (verify)
             {
-                foreach (InvPayload payload in InvPayload.CreateGroup(InventoryType.TX, context.TransactionHashes))
-                    localNode.Tell(Message.Create(MessageCommand.Inv, payload));
+                VerifyResult result = tx.Verify(neoSystem.Settings, context.Snapshot, context.VerificationContext);
+                if (result != VerifyResult.Succeed)
+                {
+                    Log($"Rejected tx: {tx.Hash}, {result}{Environment.NewLine}{tx.ToArray().ToHexString()}", LogLevel.Warning);
+                    RequestChangeView(result == VerifyResult.PolicyFail ? ChangeViewReason.TxRejectedByPolicy : ChangeViewReason.TxInvalid);
+                    return false;
+                }
             }
-            ChangeTimer(TimeSpan.FromMilliseconds((neoSystem.Settings.MillisecondsPerBlock << (context.ViewNumber + 1)) - (context.ViewNumber == 0 ? neoSystem.Settings.MillisecondsPerBlock : 0)));
+            context.Transactions[tx.Hash] = tx;
+            context.VerificationContext.AddTransaction(tx);
+            return CheckPrepareResponse();
+        }
+
+        private void ChangeTimer(TimeSpan delay)
+        {
+            clock_started = TimeProvider.Current.UtcNow;
+            expected_delay = delay;
+            timer_token.CancelIfNotNull();
+            timer_token = Context.System.Scheduler.ScheduleTellOnceCancelable(delay, Self, new Timer
+            {
+                Height = context.Block.Index,
+                ViewNumber = context.ViewNumber
+            }, ActorRefs.NoSender);
+        }
+
+        // this function increases existing timer (never decreases) with a value proportional to `maxDelayInBlockTimes`*`Blockchain.MillisecondsPerBlock`
+        private void ExtendTimerByFactor(int maxDelayInBlockTimes)
+        {
+            TimeSpan nextDelay = expected_delay - (TimeProvider.Current.UtcNow - clock_started) + TimeSpan.FromMilliseconds(maxDelayInBlockTimes * neoSystem.Settings.MillisecondsPerBlock / (double)context.M);
+            if (!context.WatchOnly && !context.ViewChanging && !context.CommitSent && (nextDelay > TimeSpan.Zero))
+                ChangeTimer(nextDelay);
+        }
+
+        protected override void PostStop()
+        {
+            Log("OnStop");
+            started = false;
+            Context.System.EventStream.Unsubscribe(Self);
+            context.Dispose();
+            base.PostStop();
+        }
+
+        public static Props Props(NeoSystem neoSystem, Settings dbftSettings, Wallet wallet)
+        {
+            return Akka.Actor.Props.Create(() => new ConsensusService(neoSystem, dbftSettings, wallet));
+        }
+
+        private static void Log(string message, LogLevel level = LogLevel.Info)
+        {
+            Utility.Log(nameof(ConsensusService), level, message);
         }
     }
 }
