+++ tests/Neo.Network.RPC.Tests/RpcTestCases.json
@@ -71,12 +71,13 @@
       "jsonrpc": "2.0",
       "id": 1,
       "result": {
-        "hash": "0x10fa4a0cb88346a00501a926cb9cbd6bf0cbd858319756c0bffcfb06e8901b92",
-        "size": 462,
+        "hash": "0x6d1556889c92249da88d2fb7729ae82fb2cc1b45dcd9030a40208b72a1d3cb83",
+        "size": 470,
         "version": 0,
         "previousblockhash": "0xaae8867e9086afaf06fd02cc538e88a69b801abd6f9d3ae39ae630e29d5b39e2",
         "merkleroot": "0xe95761f21c733ad53066786af24ee5d613b32bd5aae538df2d611492ec0cae82",
         "time": 1594867377561,
+        "nonce": "FFFFFFFFFFFFFFFF",
         "index": 7,
         "primary": 1,
         "nextconsensus": "NikvsLcNP1jWhrFPrfS3n4spEASgdNYTG2",
@@ -129,12 +130,13 @@
       "jsonrpc": "2.0",
       "id": 1,
       "result": {
-        "hash": "0x10fa4a0cb88346a00501a926cb9cbd6bf0cbd858319756c0bffcfb06e8901b92",
-        "size": 462,
+        "hash": "0x6d1556889c92249da88d2fb7729ae82fb2cc1b45dcd9030a40208b72a1d3cb83",
+        "size": 470,
         "version": 0,
         "previousblockhash": "0xaae8867e9086afaf06fd02cc538e88a69b801abd6f9d3ae39ae630e29d5b39e2",
         "merkleroot": "0xe95761f21c733ad53066786af24ee5d613b32bd5aae538df2d611492ec0cae82",
         "time": 1594867377561,
+        "nonce": "FFFFFFFFFFFFFFFF",
         "index": 7,
         "primary": 1,
         "nextconsensus": "NikvsLcNP1jWhrFPrfS3n4spEASgdNYTG2",
@@ -257,12 +259,13 @@
       "jsonrpc": "2.0",
       "id": 1,
       "result": {
-        "hash": "0x3508ada998654b2c3391462a5bfab10eff8ff7dfdb0875eccddd20c019ff9330",
-        "size": 105,
+        "hash": "0xbbf7e191d4947f8a4dc33477902dacd0b047e371a81c18a6df62fe0d541725f5",
+        "size": 113,
         "version": 0,
         "previousblockhash": "0x0000000000000000000000000000000000000000000000000000000000000000",
         "merkleroot": "0x735b68ca1bab490b8fd6166cba4c2ce76e97bedfe4d2df0c333970fc9862bb2b",
         "time": 1468595301000,
+        "nonce": "FFFFFFFFFFFFFFFF",
         "index": 0,
         "primary": 1,
         "nextconsensus": "NZs2zXSPuuv9ZF6TDGSWT1RBmE8rfGj7UW",
@@ -289,12 +292,13 @@
       "jsonrpc": "2.0",
       "id": 1,
       "result": {
-        "hash": "0x3508ada998654b2c3391462a5bfab10eff8ff7dfdb0875eccddd20c019ff9330",
-        "size": 105,
+        "hash": "0xbbf7e191d4947f8a4dc33477902dacd0b047e371a81c18a6df62fe0d541725f5",
+        "size": 113,
         "version": 0,
         "previousblockhash": "0x0000000000000000000000000000000000000000000000000000000000000000",
         "merkleroot": "0x735b68ca1bab490b8fd6166cba4c2ce76e97bedfe4d2df0c333970fc9862bb2b",
         "time": 1468595301000,
+        "nonce": "FFFFFFFFFFFFFFFF",
         "index": 0,
         "primary": 1,
         "nextconsensus": "NZs2zXSPuuv9ZF6TDGSWT1RBmE8rfGj7UW",
